Tried to break tests into files [WIP]
