+++ tests/testURLs.js
@@ -1,0 +1,16 @@
+const testURLs = [
+    {
+        url:
+      "https://stackoverflow.com/questions/50990292/"
+      + "using-octal-character-gives-warning-multi-character-character-constant",
+        textBoxQueryString: "#wmd-input",
+    },
+    {
+        url:
+      "https://serverfault.com/questions/971011/"
+      + "how-to-check-if-an-active-directory-server-is-reachable-from-an-ubuntu-apache-ph",
+        textBoxQueryString: "#wmd-input",
+    },
+];
+
+module.exports = testURLs;
